feat(view-training-by-categories): Add findWithWorkersAndTraining class method to Establishment model and refactor endpoint to use a transformer
